+++ test/benchee/formatters/markdown_integration_test.exs
@@ -16,6 +16,7 @@ defmodule Benchee.Formatters.MarkdownIntegrationTest do
         },
         time: 0.01,
         memory_time: 0.01,
+        reduction_time: 0.01,
         warmup: 0.02,
         print: [fast_warning: false],
         title: "My Benchmark",
@@ -50,6 +51,7 @@ defmodule Benchee.Formatters.MarkdownIntegrationTest do
         inputs: inputs,
         time: 0.01,
         memory_time: 0.01,
+        reduction_time: 0.01,
         warmup: 0.02,
         print: [fast_warning: false],
         formatters: [
@@ -93,6 +95,37 @@ defmodule Benchee.Formatters.MarkdownIntegrationTest do
       markdown = Benchee.Formatters.Markdown.render(result, title: :none)
       refute markdown =~ "# Benchmark"
       assert markdown =~ "Benchmark run"
+    end)
+  end
+
+  test "renders just fine for livebook" do
+    inputs = %{
+      "short_list" => Enum.to_list(1..10),
+      "long_list" => Enum.to_list(1..10_000)
+    }
+
+    map_fun = fn i -> [i, i * i] end
+
+    capture_io(fn ->
+      result = Benchee.run(
+        %{
+          "flat_map" => fn list -> Enum.flat_map(list, map_fun) end,
+          "map.flatten" => fn list ->
+            list |> Enum.map(map_fun) |> List.flatten()
+          end
+        },
+        inputs: inputs,
+        time: 0.01,
+        memory_time: 0.01,
+        warmup: 0.02,
+        print: [fast_warning: false],
+        formatters: [
+          Benchee.Formatters.Console
+        ]
+      )
+
+      markdown = Benchee.Formatters.Markdown.render(result, livebook: true)
+      assert String.starts_with?(markdown, "__Input:")
     end)
   end
 end